Move UUID implementations to other classes
