+++ rxandroidble/src/main/java/com/polidea/rxandroidble2/internal/operations/LegacyScanOperation.java
@@ -9,6 +9,7 @@ import com.polidea.rxandroidble2.LogConstants;
 import com.polidea.rxandroidble2.internal.RxBleLog;
 import com.polidea.rxandroidble2.internal.logger.LoggerUtil;
 import com.polidea.rxandroidble2.internal.scan.RxBleInternalScanResultLegacy;
+import com.polidea.rxandroidble2.internal.scan.ScanRecordImplCompat;
 import com.polidea.rxandroidble2.internal.util.RxBleAdapterWrapper;
 
 import java.util.Collections;
@@ -19,19 +20,13 @@ import java.util.UUID;
 import io.reactivex.ObservableEmitter;
 
 public class LegacyScanOperation extends ScanOperation<RxBleInternalScanResultLegacy, BluetoothAdapter.LeScanCallback> {
-
-    @SuppressWarnings("deprecation")
-    final com.polidea.rxandroidble2.internal.util.UUIDUtil uuidUtil;
     @Nullable
     final Set<UUID> filterUuids;
 
     @SuppressWarnings("deprecation")
     public LegacyScanOperation(UUID[] filterServiceUUIDs,
-                               RxBleAdapterWrapper rxBleAdapterWrapper,
-                               final com.polidea.rxandroidble2.internal.util.UUIDUtil uuidUtil) {
+                               RxBleAdapterWrapper rxBleAdapterWrapper) {
         super(rxBleAdapterWrapper);
-
-        this.uuidUtil = uuidUtil;
         if (filterServiceUUIDs != null && filterServiceUUIDs.length > 0) {
             this.filterUuids = new HashSet<>(filterServiceUUIDs.length);
             Collections.addAll(filterUuids, filterServiceUUIDs);
@@ -53,7 +48,7 @@ public class LegacyScanOperation extends ScanOperation<RxBleInternalScanResultLe
                             LoggerUtil.bytesToHex(scanRecord)
                     );
                 }
-                if (filterUuids == null || uuidUtil.extractUUIDs(scanRecord).containsAll(filterUuids)) {
+                if (filterUuids == null || ScanRecordImplCompat.extractUUIDs(scanRecord).containsAll(filterUuids)) {
                     emitter.onNext(new RxBleInternalScanResultLegacy(device, rssi, scanRecord));
                 }
             }
